Refactored Commands, Storage, UI, Parser and added packages
