+++ src/main/java/kervyn/Tasks/Deadline.java
@@ -1,4 +1,4 @@
-package Tasks;
+package kervyn.Tasks;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
