+++ src/main/java/kervyn/Tasks/Event.java
@@ -1,6 +1,5 @@
-package Tasks;
+package kervyn.Tasks;
 
-import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
