+++ src/main/java/kervyn/Tasks/TaskList.java
@@ -1,4 +1,4 @@
-package Tasks;
+package kervyn.Tasks;
 
 import java.util.ArrayList;
