+++ src/main/java/kervyn/Tasks/ToDo.java
@@ -1,4 +1,4 @@
-package Tasks;
+package kervyn.Tasks;
 public class ToDo extends Task {
     public ToDo(String description, boolean status) {
         super(description, status, Type.TODO);
