+++ src/main/java/kervyn/Tasks/Type.java
@@ -1,4 +1,4 @@
-package Tasks;
+package kervyn.Tasks;
 
 public enum Type {
         TODO,
